lkft: add support for ignore jobs failures
Sometimes, there would be problem with the squad side,
that the status is not correct, but it's not the lava job side problem.
for such case, we could specify this option to ignore the failure,
and continue to get report in time.

This is a workaround for the problem here:
    https://linaro.atlassian.net/browse/STG-3002
like this job
    https://lkft.validation.linaro.org/scheduler/job/4408903
for the build here:
    https://qa-reports.linaro.org/android-lkft/android-hikey-linaro-4.4-aosp-premerge-ci/build/4.4.300-0359ea1dbe80/testjobs/
The lava job itself is finished successfully, but there is a problem for squad
to fetch and update with the latest status
    <ProtocolError for https://qa-reports-bot:*****@lkft.validation.linaro.org//RPC2: 503 503 Server Error: Service Unavailable for url: https://qa-reports-bot:*****@lkft.validation.linaro.org//RPC2>

Signed-off-by: Yongqin Liu <yongqin.liu@linaro.org>